server/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data science libraries (use pre-built wheels on Windows)
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# ML libraries
xgboost>=2.0.0
shap>=0.43.0

# Visualization
matplotlib>=3.8.0

# Supabase and storage
supabase>=2.3.0
python-multipart>=0.0.6

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Security and auth
python-jose[cryptography]>=3.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Background tasks (optional)
celery>=5.3.0
redis>=5.0.0
